requirements.txt
aioflask==0.4.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-WTF==1.0.1
future==0.18.2
greenlet==2.0.0.post0
greenletio==0.9.0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
itchat==1.2.32
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
ntchat==0.1.16
Pillow==9.2.0
pip==22.3
PyAutoGUI==0.9.53
pyee==9.0.4
PyMsgBox==1.0.9
pypiwin32==223
pypng==0.20220715.0
PyQRCode==1.2.1
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyScreeze==0.1.28
pytweening==1.0.4
pywin32==305
PyYAML==6.0
requests==2.28.1
setuptools==65.5.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
Werkzeug==2.2.2
wheel==0.37.1
WTForms==3.0.1
xcgui==0.1.6
zipp==3.10.0
